requirements.txt
boto3
elasticsearch-curator
elasticsearch>=6.0.0,<7.0.0
requests-aws4auth
